Move Radarr history models to separate files
