+++ Cantinarr/Features/Radarr/Models/RadarrQualityDetail.swift
@@ -1,0 +1,11 @@
+// File: RadarrQualityDetail.swift
+// Purpose: Defines RadarrQualityDetail component for Cantinarr
+
+import Foundation
+
+struct RadarrQualityDetail: Codable {
+    let id: Int?
+    let name: String?
+    let source: String?
+    let resolution: Int?
+}
