+++ Cantinarr/Features/Radarr/Models/RadarrRevision.swift
@@ -1,0 +1,10 @@
+// File: RadarrRevision.swift
+// Purpose: Defines RadarrRevision component for Cantinarr
+
+import Foundation
+
+struct RadarrRevision: Codable {
+    let version: Int?
+    let real: Int?
+    let isRepack: Bool?
+}
